requirements.txt
selenium==4.29.0
beautifulsoup4==4.13.3
pandas==2.2.3
python-dotenv==0.21.0
geckodriver_autoinstaller==0.1.0
requests==2.32.3
openai==1.65.4
django==5.1.7